refactor out unused variable
