+++ internal/publish/bundle.go
@@ -1,0 +1,60 @@
+package publish
+
+// Copyright (C) 2025 by Posit Software, PBC.
+
+import (
+	"github.com/posit-dev/publisher/internal/bundles"
+	"github.com/posit-dev/publisher/internal/events"
+	"github.com/posit-dev/publisher/internal/logging"
+	"github.com/posit-dev/publisher/internal/types"
+	"io"
+	"os"
+)
+
+func (p *defaultPublisher) createBundle() (*os.File, error) {
+	manifest := bundles.NewManifestFromConfig(p.Config)
+	p.log.Debug("Built manifest from config", "config", p.ConfigName)
+
+	if p.Config.R != nil {
+		rPackages, err := p.getRPackages()
+		if err != nil {
+			return nil, err
+		}
+		manifest.Packages = rPackages
+	}
+	p.log.Debug("Generated manifest:", manifest)
+
+	// Create Bundle step
+	op := events.PublishCreateBundleOp
+	prepareLog := p.log.WithArgs(logging.LogKeyOp, op)
+
+	bundler, err := bundles.NewBundler(p.Dir, manifest, p.Config.Files, p.log)
+	if err != nil {
+		return nil, err
+	}
+
+	p.emitter.Emit(events.New(op, events.StartPhase, events.NoError, createBundleStartData{}))
+	prepareLog.Info("Preparing files")
+	bundleFile, err := os.CreateTemp("", "bundle-*.tar.gz")
+	if err != nil {
+		return nil, types.OperationError(op, err)
+	}
+	manifest, err = bundler.CreateBundle(bundleFile)
+	if err != nil {
+		return nil, types.OperationError(op, err)
+	}
+
+	_, err = bundleFile.Seek(0, io.SeekStart)
+	if err != nil {
+		return nil, types.OperationError(op, err)
+	}
+	prepareLog.Info("Done preparing files", "filename", bundleFile.Name())
+	p.emitter.Emit(events.New(op, events.SuccessPhase, events.NoError, createBundleSuccessData{
+		Filename: bundleFile.Name(),
+	}))
+
+	// Update deployment record with new information
+	p.Target.Files = manifest.GetFilenames()
+
+	return bundleFile, nil
+}
